requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
tqdm
pandas
sentence-transformers
faiss-cpu
python-dotenv
streamlit
openai
httpx
uvloop
orjson
